Split gmp code into its own file. Added path.resolve and base.path to avoid the "what context is this file load/write happening in" problem. GMP data now refreshes on bot start.
