+++ base.js
@@ -1,0 +1,3 @@
+'uses strict';
+
+exports.path = __dirname;
