+++ plugins/elite/refreshgmp
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node gmp.js
