+++ server/bot.js
@@ -8,6 +8,8 @@ var utils = require('./utils.js');
 var package_json = require("../package.json");
 var fs = require("fs");
 var _ = require("underscore");
+var base = require('../base.js');
+var path = require('path');
 var version;
 
 var config;
@@ -36,26 +38,28 @@ var messagebox;
 var aliases;
 
 try{
-    messagebox = require("../messagebox.json");
+    console.log('  - Loading ' + path.resolve(base.path, "messagebox.json"));
+    messagebox = require(path.resolve(base.path, "messagebox.json"));
 } catch(e) {
     //no stored messages
     messagebox = {};
 }
 try{
-    aliases = require("../alias.json");
+    console.log('  - Loading ' + path.resolve(base.path, "alias.json"));
+    aliases = require(path.resolve(base.path, "alias.json"));
 } catch(e) {
     //No aliases defined
     aliases = {};
 }
 
 function updateMessagebox(){
-    fs.writeFile("../messagebox.json",JSON.stringify(messagebox,null,2), null);
+    fs.writeFile(path.resolve(base.path, "messagebox.json"),JSON.stringify(messagebox,null,2), null);
 }
 
 function checkForMessages(bot, user) {
     try{
         if(messagebox.hasOwnProperty(user.id)){
-            var message = messagebox[user.id];      
+            var message = messagebox[user.id];
             var outputArray = [];
             if (message.content) {
                 outputArray.push(message.content);
@@ -124,7 +128,7 @@ function makeAliasStructFromArgs(args) {
 }
 
 function writeAliases() {
-    fs.writeFile("../alias.json",JSON.stringify(aliases,null,2), null);
+    fs.writeFile(path.resolve(base.path, "alias.json"),JSON.stringify(aliases,null,2), null);
 }
 
 function makeAliasFromArgs(args, addExtrasCallback) {
@@ -686,27 +690,14 @@ function startBot(bot, cfg) {
     }
 
     for (var i = 0; i < plugins.length; i += 1) {
-        // Prior to 2.0, the server files were in the root, with plugins a subdir off of them.
-        // In 2.0, the server files are in a server directory. This means that config files from
-        // pre-2.0 have different plugin paths than in 2.0. Rather than require the config
-        // files to be changed, we'll check one level down if the plugin can't be located...
         try {
-            plugins[i].plugin = require(plugins[i].path);
+            plugins[i].plugin = require(path.resolve(base.path, plugins[i].path));
 
             debugLog("Loaded plugin " + plugins[i].name);
         } catch(e) {
-            ; // Do nothing. We're trying again one dir down
-        }
-        if (!plugins[i].plugins) {
-            try {
-                plugins[i].plugin = require('../' + plugins[i].path);
-
-                debugLog("Loaded plugin " + plugins[i].name);
-            } catch(e) {
-                utils.logError("Couldn't load " + plugins[i].name, e);
-                if (config.DEBUG) {
-                    throw e;
-                }
+            utils.logError("Couldn't load " + plugins[i].name, e);
+            if (config.DEBUG) {
+                throw e;
             }
         }
         try {
